+++ include/render/brdf/BTDF.h
@@ -45,7 +45,7 @@ namespace render {
     virtual bool totalInternalReflection(const Rayd& ray, const HitPoint& hitPoint) const = 0;
 
     int flags() const override {
-      return BSDF::Specular | BSDF::Transmission;
+      return BSDF::Specular | BSDF::Transmission | BSDF::Delta;
     }
   };
 }